% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_epidemiology_packages}
\alias{cran_epidemiology_packages}
\title{A data.frame of historical metadata from CRAN packages epidemiology.}
\format{
A data.frame with 5 variables:
\describe{
\item{Package}{package name}
\item{Version}{package version}
\item{Authors@R}{authors as listed in the \code{Authors@R} field from the
\code{DESCRIPTION} file}
\item{Author}{authors as listed in the \code{Author} field from the
\code{DESCRIPTION} file}
\item{Maintainer}{package maintainer}
}
}
\usage{
cran_epidemiology_packages
}
\description{
A data.frame of historical metadata from CRAN packages epidemiology.
}
\keyword{datasets}
