% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_brackets.R
\name{remove_brackets}
\alias{remove_brackets}
\title{Remove content between brackets}
\usage{
remove_brackets(string, bracket = c("(", "<", "["))
}
\arguments{
\item{string}{A character vector to clean}

\item{bracket}{Type of bracket containing the content to remove.
Must be one of "(" (default), "<", "["}
}
\value{
A character vector of the same size of \code{string}, where
elements has been stripped of their \code{bracket}ted content.
}
\description{
Remove content between brackets
}
\examples{
remove_brackets(
  "R source code is versioned with svn (and mirrored on GitHub)"
)

remove_brackets(
  c("r-project.org <https://r-project.org>", "python.org <https://python.org>"),
  bracket = "<"
)

remove_brackets(
  "Content in square bracket [is removed] in the middle of a string",
  bracket = "["
)
}
