% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{authoritative-package}
\alias{authoritative}
\alias{authoritative-package}
\title{authoritative: Your Package Title in Titlecase}
\description{
Your package description. It must end with a period (".") and include relevant bibliographical references if applicable, using the following format: Author et al. (2023) \doi{10.5281/zenodo.6619350}.
}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

}
\keyword{internal}
