% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_authors_r.R
\name{parse_authors_r}
\alias{parse_authors_r}
\title{Parse the \code{Authors@R} field from a DESCRIPTION file}
\usage{
parse_authors_r(authors_r_string)
}
\arguments{
\item{authors_r_string}{A character containing the \code{Authors@R} field from a
\code{DESCRIPTION} file}
}
\value{
A \code{person} object
}
\description{
Parse the \code{Authors@R} field from a DESCRIPTION file into a \code{person} object
}
\examples{
# Read from a DESCRIPTION file directly
devtools_description <- system.file("DESCRIPTION", package = "devtools")
authors_r_devtools <- read.dcf(devtools_description, "Authors@R")

parse_authors_r(authors_r_devtools)

# Read from a database of CRAN metadata
cran_epidemiology_packages |> 
  dplyr::pull(`Authors@R`) |> 
  parse_authors_r() |> 
  head()

}
